+++ src/accounts/managers.py
@@ -5,6 +5,21 @@ from django.utils import timezone
 class UserManager(BaseUserManager):
     use_for_related_fields = True
 
+    def create_user(self, email, password, **extra_fields):
+        """
+        Create and save a user with the given email and password.
+        """
+        if not email:
+            raise ValueError("The Email must be set")
+        extra_fields.setdefault("is_staff", False)
+        extra_fields.setdefault("is_superuser", False)
+
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
+        user.set_password(password)
+        user.save()
+        return user
+
     def create_superuser(self, email, username, password, **extra_fields):
         """Creates a superuser with the given email and password."""
         now = timezone.now()