+++ src/oidc/oidc.py
@@ -1,10 +1,9 @@
 import json
 
-from allauth.socialaccount.adapter import get_adapter
-from allauth.socialaccount.providers.openid_connect.views import OpenIDConnectOAuth2Adapter
-from allauth.utils import build_absolute_uri
 from django.core.exceptions import PermissionDenied
-from django.urls import reverse
+from mozilla_django_oidc.auth import OIDCAuthenticationBackend
+
+from .models import OidcLogin, ProviderChoice
 
 MANAGER_ID = 4
 ADMIN_ID = 2
@@ -17,25 +16,34 @@ ALLOWED_PROFILES = {MANAGER_ID, ADMIN_ID, GUN_ID, GUN_READER_ID}
 ALLOWED_PERIMETER = "ICPE"
 
 
-class MonAiotDConnectAdapter(OpenIDConnectOAuth2Adapter):
-    def complete_login(self, request, app, token, **kwargs):
-        """Ensure userinfo carries relevant permissions to log user in or create their account"""
+class MonAiotOidcBackend(OIDCAuthenticationBackend):
+    def create_user(self, claims):
+        """Return object for a newly created user account."""
+        email = claims.get("email")
 
-        response = (
-            get_adapter()
-            .get_requests_session()
-            .get(self.profile_url, headers={"Authorization": "Bearer " + str(token.token)})
-        )
-        response.raise_for_status()
-        extra_data = response.json()
-        droits_data = extra_data.get("droits", None)
+        user = self.UserModel.objects.create_user(username=email, email=email, password="", monaiot_signup=True)
+
+        OidcLogin.objects.create(user=user, provider=ProviderChoice.MONAIOT, info=claims, account_created=True)
+        return user
+
+    def update_user(self, user, claims):
+        user = super().update_user(user, claims)
+        if not user.monaiot_connexion:
+            user.monaiot_connexion = True
+            user.save()
+
+        OidcLogin.objects.create(user=user, info=claims, provider=ProviderChoice.MONAIOT)
+        return user
+
+    def verify_claims(self, claims):
+        access_granted = False
+        verified = super().verify_claims(claims)
+        droits_data = claims.get("droits", None)
 
         if not droits_data:
             raise PermissionDenied
 
         droits = json.loads(droits_data)
-
-        access_granted = False
         for droit in droits:
             id_profil = droit.get("id_profil")
             perimetre_ic = droit.get("perimetre_ic")
@@ -53,16 +61,6 @@ class MonAiotDConnectAdapter(OpenIDConnectOAuth2Adapter):
                     access_granted = False
                 if id_nature_service != GUN_READER_ALLOWED_SERVICE_ID:
                     access_granted = False
-
         if not access_granted:
             raise PermissionDenied
-
-        # Gun readers require extra verification
-
-        return self.get_provider().sociallogin_from_response(request, extra_data)
-
-    def get_callback_url(self, request, app):
-        callback_url = reverse("monaiot_callback", kwargs={"provider_id": self.provider_id})
-
-        protocol = self.redirect_uri_protocol
-        return build_absolute_uri(request, callback_url, protocol)
+        return verified