+++ changelog.md
@@ -5,6 +5,10 @@ Les changements importants de Trackdéchets préparation inspection sont documen
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et le projet suit un schéma de versioning inspiré de [Calendar Versioning](https://calver.org/).
 
+## 24/02/2025
+
+- Utilisation de mozilla-oidc pour la connexion MonAIOT
+
 ## 20/02/2025
 
 - Suppression de la variable d'environnement permettant de sélectionner la query utilisée pour la recherche de bsds